requirements.txt
uagents
mcp
openai
python-dotenv
httpx
stripe
